style: UneditabelMemo 코드 분리 및 정리
