+++ src/pages/home/subPages/components/memo/UneditableMemo/MemoFooter/index.ts
@@ -1,0 +1,1 @@
+export { default as MemoFooter } from './MemoFooter';
