+++ src/pages/home/subPages/components/memo/UneditableMemo/MemoHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as MemoHeader } from './MemoHeader';
